Drop some classes not needed by the HTML checker
This change removes some classes that aren’t necessary in the context of
using the CSS-checking backend in the context of the HTML checker.
Dropping these classes eliminates the need for any additional
dependencies other than Jigsaw.

This change also adjusts some remaining code to remove dependency on a
couple of the deleted classes.